+++ R/datadocs.R
@@ -204,8 +204,13 @@ NULL
 
 #' The currently active version of ICD-10-CM in this package.
 #'
-#' The default is the latest version, but it may be set using
-#' `set_icd10cm_ver()`
+#' This is work in progress, so not exported. The default is the latest version,
+#' however, depending on the context, it seems not to respect the R option, e.g.
+#' during testing. `set_icd10cm_ver()` may be used to set the active version,
+#' and `get_icd10cm_ver()` gives the current setting. `get_icd10cm_version(ver =
+#' "2015")` will return the 2015 version. `get_icd10cm_active()` will return the
+#' currently active version, as set by the option, then default being the most
+#' recent available.
 #' @examples
 #' \dontrun{
 #' set_icd10cm_active_ver("2016")
@@ -213,8 +218,8 @@ NULL
 #' @details This is an active binding, so is exported explicitly
 #' @docType data
 #' @keywords datasets
+#' @seealso get_icd10cm_version get_icd10cm_ver set_icd10cm_ver
 #' @name icd10cm_active
-#' @export
 NULL
 
 #' French ICD-10-FR modification of WHO ICD-10 used in France
@@ -344,6 +349,7 @@ NULL
 #' @source
 #' \url{http://www.cms.gov/Medicare/Coding/ICD9ProviderDiagnosticCodes/codes.html}
 #' @name icd9cm_leaf_v32
+#' @export
 NULL
 
 #' Deprecated ICD-9-CM leaf/billable codes